AMP-2401
git-svn-id: file:///svn/trunk/amp@8464 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/dgfoundation/amp/ecs/common/ECSParameters.java
@@ -12,14 +12,13 @@ public class ECSParameters implements Serializable{
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private int synchronizeDelay;
+	private long synchronizeDelay;
 	private boolean runOnceCustom;
 	
-	
-	public int getSynchronizeDelay() {
+	public long getSynchronizeDelay() {
 		return synchronizeDelay;
 	}
-	public void setSynchronizeDelay(int synchronizeDelay) {
+	public void setSynchronizeDelay(long synchronizeDelay) {
 		this.synchronizeDelay = synchronizeDelay;
 	}
 	public boolean isRunOnceCustom() {
